+++ webrtc/WebRtcSignalingSession.cpp
@@ -13,7 +13,6 @@
 #include "WebRtcTransport.h"
 #include "Util/util.h"
 #include "Common/config.h"
-#include "server/WebApi.h"
 
 using namespace std;
